#2 Role Admin

+++ resources/views/categories/index.blade.php
@@ -2,22 +2,26 @@
 
 @section('content')
     <div class="container">
+        @if(\Illuminate\Support\Facades\Auth::user()->is_admin === 1)
         <div class="d-flex justify-content-end mb-2">
             <a href=" {{ route('categories.create') }}" class="btn btn-primary"> Ajouter une catégorie</a>
         </div>
+        @endif
 
         <div class="row row-cols-1 row-cols-md-3 g-4">
             @foreach($categories as $category)
-                <div class="col py-4">
+                <div class="col">
                     <div class="card h-100 shadow-sm">
                         <img class="card-img-top" src="{{URL::asset('/img/test.png')}}" alt="Card image cap">
                         <div class="card-body">
                             <h5 class="card-title">{{ $category->name }}</h5>
                         </div>
-                        <a href="{{route('categories.edit', $category->id)}}" class="btn btn-info btn-sm"> Modifier</a>
-                        <button class="btn btn-danger btn-sm" onclick="handleDelete({{ $category->id }})">
-                            Supprimer
-                        </button>
+                        @if(\Illuminate\Support\Facades\Auth::user()->is_admin === 1)
+                            <a href="{{route('categories.edit', $category->id)}}" class="btn btn-info btn-sm"> Modifier</a>
+                            <button class="btn btn-danger btn-sm" onclick="handleDelete({{ $category->id }})">
+                                Supprimer
+                            </button>
+                        @endif
                     </div>
                 </div>
             @endforeach
